requirements.txt
asgiref==3.4.1
Django==3.2.10
django-rest-framework==0.1.0
djangorestframework==3.12.4
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
